Imported Upstream version 2.90

+++ src/locking.c
@@ -6,7 +6,7 @@
  ************************************************************************/
 #ifdef RCS
 static /*const*/char rcsid[]=
- "$Id: locking.c,v 1.13 1993/02/11 12:08:34 berg Exp $";
+ "$Id: locking.c,v 1.16 1993/06/21 14:24:30 berg Exp $";
 #endif
 #include "procmail.h"
 #include "robust.h"
@@ -24,12 +24,18 @@ void lockit(name,lockp)char*name;char**const lockp;
    }				  /* to prevent deadlocks (I hate deadlocks) */
   if(!*name)
      return;
-  name=tstrdup(name); /* allocate now, so we won't hang on memory *and* lock */
   if(!strcmp(name,defdeflock))	       /* is it the system mailbox lockfile? */
-     setgid(sgid);		       /* try and get some extra permissions */
+#ifdef fdlock
+     if(!accspooldir)
+      { yell("Bypassed locking",name);return;
+      }
+     else
+#endif
+	setgid(sgid);		       /* try and get some extra permissions */
+  name=tstrdup(name); /* allocate now, so we won't hang on memory *and* lock */
   for(lcking|=lck_LOCKFILE;;)
    { yell("Locking",name);	    /* in order to cater for clock skew: get */
-     if(!xcreat(name,LOCKperm,&t,(int*)0))     /* time t from the filesystem */
+     if(!xcreat(name,LOCKperm,&t,0))	       /* time t from the filesystem */
       { *lockp=name;break;			   /* lock acquired, hurray! */
       }
      switch(errno)
@@ -96,9 +102,12 @@ void lcllock P((void))				    /* lock a local lockfile */
 void unlock(lockp)char**const lockp;
 { lcking|=lck_LOCKFILE;
   if(*lockp)
-   { yell("Unlocking",*lockp);
-     if(!strcmp(*lockp,defdeflock))    /* is it the system mailbox lockfile? */
-	setgid(sgid);		       /* try and get some extra permissions */
+   { if(!strcmp(*lockp,defdeflock))    /* is it the system mailbox lockfile? */
+	if(!accspooldir)
+	   goto no_lock;
+	else
+	   setgid(sgid);	       /* try and get some extra permissions */
+     yell("Unlocking",*lockp);
      if(unlink(*lockp))
 	nlog("Couldn't unlock"),logqnl(*lockp);
      if(rc!=rc_INIT)				     /* we opened any rcfile */
@@ -107,24 +116,27 @@ void unlock(lockp)char**const lockp;
 	free(*lockp);
      *lockp=0;
    }
+no_lock:
   lcking&=~lck_LOCKFILE;
   if(nextexit==1)	    /* make sure we are not inside terminate already */
      elog(newline),terminate();
 }
 					/* an NFS secure exclusive file open */
-xcreat(name,mode,tim,chowned)const char*const name;const mode_t mode;
- time_t*const tim;int*const chowned;
+xcreat(name,mode,tim,chownit)const char*const name;const mode_t mode;
+ time_t*const tim;const int chownit;
 { char*p;int j= -2,i;
   i=lastdirsep(name)-name;strncpy(p=malloc(i+UNIQnamelen),name,i);
   if(unique(p,p+i,mode,verbose))       /* try and rename the unique filename */
-   { if(chowned)
-	*chowned=chown(p,uid,sgid);			 /* try and chown it */
+   { if(chownit&&chown(p,uid,sgid))			 /* try and chown it */
+      { unlink(p);goto sorry;			 /* forget it, no permission */
+      }
      if(tim)
       { struct stat stbuf;	 /* return the filesystem time to the caller */
 	stat(p,&stbuf);*tim=stbuf.st_mtime;
       }
      j=myrename(p,name);
    }
+sorry:
   free(p);return j;
 }
 	/* if you've ever wondered what conditional compilation was good for */
@@ -145,7 +157,7 @@ static struct flock flck;		/* why can't it be a local variable? */
 #define REITfcntl	0
 #endif /* NOfcntl_lock */
 #ifdef USElockf
-static long oldlockoffset;
+static off_t oldlockoffset;
 #define REITlockf	1
 #else
 #define REITlockf	0
@@ -168,7 +180,7 @@ fdlock(fd)
 #ifndef NOfcntl_lock
      ret=fcntl(fd,F_SETLKW,&flck);
 #ifdef USElockf
-     if((ret|=lockf(fd,F_TLOCK,0L))&&(errno==EAGAIN||errno==EACCES||
+     if((ret|=lockf(fd,F_TLOCK,(off_t)0))&&(errno==EAGAIN||errno==EACCES||
       errno==EWOULDBLOCK))
 ufcntl:
       { flck.l_type=F_UNLCK;fcntl(fd,F_SETLK,&flck);continue;
@@ -176,7 +188,7 @@ ufcntl:
 #ifdef USEflock
      if((ret|=flock(fd,LOCK_EX|LOCK_NB))&&(errno==EAGAIN||errno==EACCES||
       errno==EWOULDBLOCK))
-      { lockf(fd,F_ULOCK,0L);goto ufcntl;
+      { lockf(fd,F_ULOCK,(off_t)0);goto ufcntl;
       }
 #endif /* USEflock */
 #else /* USElockf */
@@ -189,11 +201,11 @@ ufcntl:
 #endif /* USElockf */
 #else /* NOfcntl_lock */
 #ifdef USElockf
-     ret=lockf(fd,F_LOCK,0L);
+     ret=lockf(fd,F_LOCK,(off_t)0);
 #ifdef USEflock
      if((ret|=flock(fd,LOCK_EX|LOCK_NB))&&(errno==EAGAIN||errno==EACCES||
       errno==EWOULDBLOCK))
-      { lockf(fd,F_ULOCK,0L);continue;
+      { lockf(fd,F_ULOCK,(off_t)0);continue;
       }
 #endif /* USEflock */
 #else /* USElockf */
@@ -215,7 +227,7 @@ fdunlock P((void))
   i|=flock(oldfdlock,LOCK_UN);
 #endif
 #ifdef USElockf
-  lseek(oldfdlock,oldlockoffset,SEEK_SET);i|=lockf(oldfdlock,F_ULOCK,0L);
+  lseek(oldfdlock,oldlockoffset,SEEK_SET);i|=lockf(oldfdlock,F_ULOCK,(off_t)0);
 #endif
 #ifndef NOfcntl_lock
   flck.l_type=F_UNLCK;i|=fcntl(oldfdlock,F_SETLK,&flck);
